refactor & restructuring data_access and use_case
