+++ src/main/java/data_access/fileuser/FileUserDataAccessObject.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.fileuser;
 
 import domains.user.entity.User;
 import domains.user.entity.UserFactory;
