+++ src/main/java/data_access/inmemoryuser/InMemoryUserDataAccessObject.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.inmemoryuser;
 
 public class InMemoryUserDataAccessObject {
 }
